Refactoring & added tests
