+++ autode/ext/src/utils.cpp
