+++ autode/ext/wrappers.pxd
@@ -3,7 +3,7 @@ from libcpp.vector cimport vector
 from libcpp cimport bool as bool_t
 
 
-cdef extern from "molecule.h" namespace "autode":
+cdef extern from "include/molecule.h" namespace "autode":
     cdef cppclass Molecule:
         # Constructors
         Molecule()
@@ -18,12 +18,12 @@ cdef extern from "molecule.h" namespace "autode":
         void rotate_dihedrals()
 
 
-cdef extern from "potentials.h" namespace "autode":
+cdef extern from "include/potentials.h" namespace "autode":
     cdef cppclass Potential:
         pass
 
 
-cdef extern from "potentials.h" namespace "autode":
+cdef extern from "include/potentials.h" namespace "autode":
     cdef cppclass RBPotential(Potential):
         RBPotential()
         RBPotential(int,             # Exponent
@@ -33,14 +33,14 @@ cdef extern from "potentials.h" namespace "autode":
                     vector[double])  # Pairwise repulsion coefficients
 
 
-cdef extern from "potentials.h" namespace "autode":
+cdef extern from "include/potentials.h" namespace "autode":
     cdef cppclass RDihedralPotential(Potential):
         RDihedralPotential()
         RDihedralPotential(int,             # Exponent
                            vector[bool_t])  # Pairs that should be considered
 
 
-cdef extern from "potentials.h" namespace "autode":
+cdef extern from "include/potentials.h" namespace "autode":
     cdef cppclass RRingDihedralPotential(Potential):
         RRingDihedralPotential()
         RRingDihedralPotential(int,            # Exponent
@@ -49,7 +49,7 @@ cdef extern from "potentials.h" namespace "autode":
                                double)         # Distance for ring closure
 
 
-cdef extern from "optimisers.h" namespace "autode":
+cdef extern from "include/optimisers.h" namespace "autode":
     cdef cppclass SDOptimiser:
         SDOptimiser()
 
@@ -61,12 +61,12 @@ cdef extern from "optimisers.h" namespace "autode":
                  double)     # Initial step size
 
 
-cdef extern from "optimisers.h" namespace "autode":
+cdef extern from "include/optimisers.h" namespace "autode":
     cdef cppclass SDDihedralOptimiser(SDOptimiser):
         SDDihedralOptimiser()
 
 
-cdef extern from "optimisers.h" namespace "autode":
+cdef extern from "include/optimisers.h" namespace "autode":
     cdef cppclass GridDihedralOptimiser:
         GridDihedralOptimiser()
 
@@ -77,7 +77,7 @@ cdef extern from "optimisers.h" namespace "autode":
                  double)  # Initial step size for SD minimisation
 
 
-cdef extern from "optimisers.h" namespace "autode":
+cdef extern from "include/optimisers.h" namespace "autode":
     cdef cppclass SGlobalDihedralOptimiser:
         SGlobalDihedralOptimiser()
 
@@ -87,7 +87,7 @@ cdef extern from "optimisers.h" namespace "autode":
                  double,     # Final tolerance on SD minimisation
                  double)     # Initial step size for SD minimisation
 
-cdef extern from "dihedrals.h" namespace "autode":
+cdef extern from "include/dihedrals.h" namespace "autode":
     cdef cppclass Dihedral:
         Dihedral()
         Dihedral(double,           # Angle
