+++ setup.py
@@ -5,12 +5,14 @@ from setuptools.extension import Extension
 extensions = [Extension('cconf_gen', ['autode/conformers/cconf_gen.pyx']),
               Extension('ade_dihedrals',
                         sources=['autode/ext/ade_dihedrals.pyx'],
+                        include_dirs=['autode/ext/include'],
                         language='c++',
                         extra_compile_args=["-std=c++11"],
                         extra_link_args=["-std=c++11"]),
 
               Extension('ade_rb_opt',
                         sources=['autode/ext/ade_rb_opt.pyx'],
+                        include_dirs=['autode/ext/include'],
                         language='c++',
                         extra_compile_args=["-std=c++11"],
                         extra_link_args=["-std=c++11"])]
